+++ config/historytrack.php
@@ -46,7 +46,7 @@ return [
      * This is the name of the table that will be created by the migration and
      * used by the HistoryTracking model shipped with this package.
      */
-    'table_name' => 'activity_log',
+    'table_name' => 'history_tracking',
 
     /*
      * This is the database connection that will be used by the migration and