+++ src/Models/HistoryTracking.php
@@ -49,11 +49,11 @@ class HistoryTracking extends Model implements ActivityContract
     public function __construct(array $attributes = [])
     {
         if (! isset($this->connection)) {
-            $this->setConnection(config('activitylog.database_connection'));
+            $this->setConnection(config('historytrack.database_connection'));
         }
 
         if (! isset($this->table)) {
-            $this->setTable(config('activitylog.table_name'));
+            $this->setTable(config('historytrack.table_name'));
         }
 
         parent::__construct($attributes);
@@ -61,7 +61,7 @@ class HistoryTracking extends Model implements ActivityContract
 
     public function subject(): MorphTo
     {
-        if (config('activitylog.subject_returns_soft_deleted_models')) {
+        if (config('historytrack.subject_returns_soft_deleted_models')) {
             return $this->morphTo()->withTrashed();
         }