+++ tests/CustomDatabaseConnectionActivityModelTest.php
@@ -6,7 +6,7 @@ use Jobful\HistoryTracking\Test\Models\CustomDatabaseConnectionOnHistoryTracking
 it('uses the database connection from the configuration', function () {
     $model = new HistoryTracking();
 
-    expect(config('activitylog.database_connection'))->toEqual($model->getConnectionName());
+    expect(config('historytrack.database_connection'))->toEqual($model->getConnectionName());
 });
 
 it('uses a custom database connection', function () {
@@ -14,12 +14,12 @@ it('uses a custom database connection', function () {
 
     $model->setConnection('custom_sqlite');
 
-    $this->assertNotEquals($model->getConnectionName(), config('activitylog.database_connection'));
+    $this->assertNotEquals($model->getConnectionName(), config('historytrack.database_connection'));
     expect('custom_sqlite')->toEqual($model->getConnectionName());
 });
 
 it('uses the default database connection when the one from configuration is null', function () {
-    app()['config']->set('activitylog.database_connection', null);
+    app()['config']->set('historytrack.database_connection', null);
 
     $model = new HistoryTracking();
 
@@ -29,6 +29,6 @@ it('uses the default database connection when the one from configuration is null
 it('uses the database connection from model', function () {
     $model = new CustomDatabaseConnectionOnHistoryTrackingModel();
 
-    $this->assertNotEquals($model->getConnectionName(), config('activitylog.database_connection'));
+    $this->assertNotEquals($model->getConnectionName(), config('historytrack.database_connection'));
     expect('custom_connection_name')->toEqual($model->getConnectionName());
 });